% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/andurinhaData.R
\docType{data}
\name{andurinhaData}
\alias{andurinhaData}
\title{Andurinha data set}
\format{
A data frame with 1736 observations of 3 peat samples.
\itemize{
 \item \emph{WN}: wave numbers of the spectra.
 \item \emph{A-C}: absorbances of a FTIR spectra of three peat samples.
}
}
\source{
EcoPast research group, Universidade de Santiago de Compostela
  (Spain) \url{https://ecopast.es}
}
\usage{
andurinhaData
}
\description{
A collection of FTIR-ATR mid-infrared spectra of peat samples
  with increasing degree of peat humification.
}
\seealso{
\code{\link{importSpectra}}, \code{\link{findPeaks}},
  \code{\link{plotPeaks}} and \code{\link{gOverview}}
}
\keyword{datasets}
